refactor(DTFS2-6935): split pagination component test helper functions into separate files
