+++ trade-finance-manager-ui/component-tests/test-assertions/pagination/index.js
@@ -1,0 +1,28 @@
+const itShouldRenderOuterDivAndDisplayTotalItems = require('./itShouldRenderOuterDivAndDisplayTotalItems');
+const itShouldRenderNavigationList = require('./itShouldRenderNavigationList');
+const itShouldNotRenderNavigationList = require('./itShouldNotRenderNavigationList');
+const itShouldRenderFirstAndPreviousLinks = require('./itShouldRenderFirstAndPreviousLinks');
+const itShouldNotRenderFirstAndPreviousLinks = require('./itShouldNotRenderFirstAndPreviousLinks');
+const itShouldRenderEllipsisAfterFirstAndPreviousLinks = require('./itShouldRenderEllipsisAfterFirstAndPreviousLinks');
+const itShouldNotRenderEllipsisAfterFirstAndPreviousLinks = require('./itShouldNotRenderEllipsisAfterFirstAndPreviousLinks');
+const itShouldRenderLinksToPagesInRange = require('./itShouldRenderLinksToPagesInRange');
+const itShouldRenderEllipsisBeforeNextAndLastLinks = require('./itShouldRenderEllipsisBeforeNextAndLastLinks');
+const itShouldNotRenderEllipsisBeforeNextAndLastLinks = require('./itShouldNotRenderEllipsisBeforeNextAndLastLinks');
+const itShouldRenderNextAndLastLinks = require('./itShouldRenderNextAndLastLinks');
+const itShouldNotRenderNextAndLastLinks = require('./itShouldNotRenderNextAndLastLinks');
+
+
+module.exports = {
+  itShouldRenderOuterDivAndDisplayTotalItems,
+  itShouldRenderNavigationList,
+  itShouldNotRenderNavigationList,
+  itShouldRenderFirstAndPreviousLinks,
+  itShouldNotRenderFirstAndPreviousLinks,
+  itShouldRenderEllipsisAfterFirstAndPreviousLinks,
+  itShouldNotRenderEllipsisAfterFirstAndPreviousLinks,
+  itShouldRenderLinksToPagesInRange,
+  itShouldRenderEllipsisBeforeNextAndLastLinks,
+  itShouldNotRenderEllipsisBeforeNextAndLastLinks,
+  itShouldRenderNextAndLastLinks,
+  itShouldNotRenderNextAndLastLinks,
+};
