+++ trade-finance-manager-ui/component-tests/test-assertions/pagination/itShouldNotRenderEllipsisAfterFirstAndPreviousLinks.js
@@ -1,0 +1,7 @@
+module.exports = (getWrapper) => {
+    it('should NOT render an ellipsis after the \'First\' and \'Previous\' links', () => {
+      const wrapper = getWrapper();
+
+      wrapper.expectElement('[data-cy="firstPreviousEllipsis"]').notToExist();
+    });
+}
