+++ trade-finance-manager-ui/component-tests/test-assertions/pagination/itShouldRenderEllipsisAfterFirstAndPreviousLinks.js
@@ -1,0 +1,9 @@
+module.exports = (getWrapper) => {
+    it('should render an ellipsis after the \'First\' and \'Previous\' links', () => {
+      const wrapper = getWrapper();
+
+      wrapper.expectElement('[data-cy="firstPreviousEllipsis"]').hasClass('govuk-body');
+      wrapper.expectElement('[data-cy="firstPreviousEllipsis"]').hasClass('govuk-!-margin-bottom-3');
+      wrapper.expectText('[data-cy="firstPreviousEllipsis"]').toRead('...');
+    });
+}
